% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bears.R
\docType{data}
\name{bears}
\alias{bears}
\title{GPS location data of bears}
\format{
A data frame with x rows and y column(s)
\describe{
  \item{column1}{Describe column here . . .}
  \item{column2}{Describe column here . . .}
 ...
}
}
\source{
\url{http://www.whereyougotthedata.com/"}
}
\usage{
bears
}
\description{
Short description of data . . .
}
\keyword{datasets}
