% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolf.R
\docType{data}
\name{wolf}
\alias{wolf}
\title{GPS location data of wolf individual}
\format{
A data frame with 8 rows and 1134 column(s)
\describe{
  \item{Object_ID}{The ID of the monitored individual.}
  \item{Collar_ID}{The ID of the collar fit on the individual.}
  \item{PubName}{The ID of the individual within the pack.}
  \item{Sex}{The sex of the individual.}
  \item{GMT_date}{Date and time in GMT.}
  \item{LMT_date}{Date and time in LMT.}
  \item{Longitude}{Northing position in WGS84.}
  \item{Latitude}{Easting position in WGS84.}
  \item{Height}{Height of position.}

}
}
\source{
Scandinavian Wolf Research Project (SKANDULV) - Contact: Camilla Wikenros
}
\usage{
wolf
}
\description{
Directly exported GPS data for the male wolf in Prästskogen territory (Sweden) during winter 2022. The intensive period went from 28-02-2022 until 18-04-2022.
}
\keyword{datasets}
