% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bears.R
\docType{data}
\name{bears}
\alias{bears}
\title{GPS location data of three bears}
\format{
A data frame with 8 rows and 2204 column
\describe{
  \item{Object_ID}{The ID of the monitored individuals.}
  \item{Collar_ID}{The ID of the collar fit on the individuals.}
  \item{PubName}{The ID of the individual within the pack.}
  \item{GMT_date}{Date and time in GMT.}
  \item{LMT_date}{Date and time in LMT.}
  \item{Longitude}{Northing position in WGS84.}
  \item{Latitude}{Easting position in WGS84.}
  \item{Height}{Height of position.}

}
}
\source{
Scandinavian Brown Bear Project - Contact: Aimee Tallian
}
\usage{
bears
}
\description{
Historic data for 3 bears(Läfmamäck = adult male; Ruta = adult female; Järpa = subadult female) from 2014 . The intensive period went from 01-05-2014 until 31-05-2014.
}
\keyword{datasets}
